fix: turn-aware context windowing prevents Gemini ordering errors
When recentMessages truncated mid-tool-chain, the first loaded message
could be an assistant tool-call instead of a user message. Gemini
rejects this with "function call turn must come after user/function
response turn."

Replace flat message-count truncation with turn-aware context windowing:
- buildContextWindow groups messages into complete conversation turns
  (user message + all following assistant/tool messages)
- Selects the most recent turns within a 30K token budget
- Never breaks mid-tool-chain, always starts at a user message
- Semantic search recovers relevant older context

Also increases recentMessages from 30 to 100 so the contextHandler
receives enough data to make smart truncation decisions.

Extracts all message-processing functions (mergeConsecutiveSameRole,
stripOrphanedToolCalls, stripImagesFromOlderMessages, buildContextWindow)
to contextWindow.ts to keep coach.ts under the 400-line cap.
